+++ client/src/components/Frame.js
@@ -6,8 +6,8 @@ const Frame = (props) => {
     <div>
       <div className={styles.bar}></div>
       <img className={[styles.logo, styles.center].join(' ')} src="/logo.png" alt="pup taguig logo" />
-      <h1 className={styles.center}>{props.h1}</h1>
-      <h2 className={styles.center}>{props.h2}</h2>
+      <h1 className={[styles.title, styles.center].join(' ')}>{props.h1}</h1>
+      <h2 className={[styles.subtitle, styles.center].join(' ')}>{props.h2}</h2>
     </div>
   );
 }